requirements.txt
Django==4.0
psycopg2_binary==2.9.6
djangorestframework==3.14.0
django-filter==23.1
drf-nested-routers==0.93.4
pytest-django==4.5.2